testsuite/pyproject.toml
[tool.poetry]
name = "velor_testsuite"
version = "0.1.0"
description = "Velor E2E tests and utilities in python, including Forge"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
click = "^8.1.3"
psutil = "^5.9.1"
requests = "^2.31.0"
pyyaml = "^6.0"
tabulate = "^0.9.0"
dacite = "^1.8.1"
kubernetes = "^26.1.0"
kubernetes-stubs = "^22.6.0.post1"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pyright = "^1.1.309"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
reportUndefinedVariable = "none"
typeCheckingMode = "basic" # TODO(rustielin): eventually make this strict
